// lib/blob/generated/parameters.ts
export const blobCacheControl: msRest.OperationParameter = {
  parameterPath: ["options", "blobHTTPHeaders", "blobCacheControl"],
  mapper: {
    serializedName: "x-ms-blob-cache-control",
    type: {
      name: "String",
    },
  },
};
export const blobContentDisposition: msRest.OperationParameter = {
  parameterPath: ["options", "blobHTTPHeaders", "blobContentDisposition"],
  mapper: {
    serializedName: "x-ms-blob-content-disposition",
    type: {
      name: "String",
    },
  },
};
export const blobContentEncoding: msRest.OperationParameter = {
  parameterPath: ["options", "blobHTTPHeaders", "blobContentEncoding"],
  mapper: {
    serializedName: "x-ms-blob-content-encoding",
    type: {
      name: "String",
    },
  },
};
export const blobContentLanguage: msRest.OperationParameter = {
  parameterPath: ["options", "blobHTTPHeaders", "blobContentLanguage"],
  mapper: {
    serializedName: "x-ms-blob-content-language",
    type: {
      name: "String",
    },
  },
};
export const blobContentLength: msRest.OperationParameter = {
  parameterPath: "blobContentLength",
  mapper: {
    required: true,
    serializedName: "x-ms-blob-content-length",
    type: {
      name: "Number",
    },
  },
};
export const blobContentMD5: msRest.OperationParameter = {
  parameterPath: ["options", "blobHTTPHeaders", "blobContentMD5"],
  mapper: {
    serializedName: "x-ms-blob-content-md5",
    type: {
      name: "ByteArray",
    },
  },
};
export const blobContentType: msRest.OperationParameter = {
  parameterPath: ["options", "blobHTTPHeaders", "blobContentType"],
  mapper: {
    serializedName: "x-ms-blob-content-type",
    type: {
      name: "String",
    },
  },
};
// ...
export const comp15: msRest.OperationQueryParameter = {
  parameterPath: "comp",
  mapper: {
    required: true,
    isConstant: true,
    serializedName: "comp",
    defaultValue: "blocklist",
    type: {
      name: "String",
    },
  },
};
export const comp2: msRest.OperationQueryParameter = {
  parameterPath: "comp",
  mapper: {
    required: true,
    isConstant: true,
    serializedName: "comp",
    defaultValue: "list",
    type: {
      name: "String",
    },
  },
};
// ...
export const ifMatch: msRest.OperationParameter = {
  parameterPath: ["options", "modifiedAccessConditions", "ifMatch"],
  mapper: {
    serializedName: "If-Match",
    type: {
      name: "String",
    },
  },
};
export const ifModifiedSince: msRest.OperationParameter = {
  parameterPath: ["options", "modifiedAccessConditions", "ifModifiedSince"],
  mapper: {
    serializedName: "If-Modified-Since",
    type: {
      name: "DateTimeRfc1123",
    },
  },
};
export const ifNoneMatch: msRest.OperationParameter = {
  parameterPath: ["options", "modifiedAccessConditions", "ifNoneMatch"],
  mapper: {
    serializedName: "If-None-Match",
    type: {
      name: "String",
    },
  },
};
// ...
export const ifUnmodifiedSince: msRest.OperationParameter = {
  parameterPath: ["options", "modifiedAccessConditions", "ifUnmodifiedSince"],
  mapper: {
    serializedName: "If-Unmodified-Since",
    type: {
      name: "DateTimeRfc1123",
    },
  },
};
export const include0: msRest.OperationQueryParameter = {
  parameterPath: ["options", "include"],
  mapper: {
    serializedName: "include",
    type: {
      name: "Enum",
      allowedValues: ["metadata"],
    },
  },
};
// ...
export const leaseId0: msRest.OperationParameter = {
  parameterPath: ["options", "leaseAccessConditions", "leaseId"],
  mapper: {
    serializedName: "x-ms-lease-id",
    type: {
      name: "String",
    },
  },
};
// ...
export const marker: msRest.OperationQueryParameter = {
  parameterPath: ["options", "marker"],
  mapper: {
    serializedName: "marker",
    type: {
      name: "String",
    },
  },
};
export const maxresults: msRest.OperationQueryParameter = {
  parameterPath: ["options", "maxresults"],
  mapper: {
    serializedName: "maxresults",
    constraints: {
      InclusiveMinimum: 1,
    },
    type: {
      name: "Number",
    },
  },
};
export const maxSize: msRest.OperationParameter = {
  parameterPath: ["options", "appendPositionAccessConditions", "maxSize"],
  mapper: {
    serializedName: "x-ms-blob-condition-maxsize",
    type: {
      name: "Number",
    },
  },
};
export const metadata: msRest.OperationParameter = {
  parameterPath: ["options", "metadata"],
  mapper: {
    serializedName: "x-ms-meta",
    type: {
      name: "Dictionary",
      value: {
        type: {
          name: "String",
        },
      },
    },
    headerCollectionPrefix: "x-ms-meta-",
  },
};
// ...
export const prefix: msRest.OperationQueryParameter = {
  parameterPath: ["options", "prefix"],
  mapper: {
    serializedName: "prefix",
    type: {
      name: "String",
    },
  },
};
// ...
export const requestId: msRest.OperationParameter = {
  parameterPath: ["options", "requestId"],
  mapper: {
    serializedName: "x-ms-client-request-id",
    type: {
      name: "String",
    },
  },
};
// ...
export const timeout: msRest.OperationQueryParameter = {
  parameterPath: ["options", "timeout"],
  mapper: {
    serializedName: "timeout",
    constraints: {
      InclusiveMinimum: 0,
    },
    type: {
      name: "Number",
    },
  },
};
export const transactionalContentMD5: msRest.OperationParameter = {
  parameterPath: ["options", "transactionalContentMD5"],
  mapper: {
    serializedName: "Content-MD5",
    type: {
      name: "ByteArray",
    },
  },
};
export const url: msRest.OperationURLParameter = {
  parameterPath: "url",
  mapper: {
    required: true,
    serializedName: "url",
    defaultValue: "",
    type: {
      name: "String",
    },
  },
  skipEncoding: true,
};
export const version: msRest.OperationParameter = {
  parameterPath: "version",
  mapper: {
    required: true,
    serializedName: "x-ms-version",
    type: {
      name: "String",
    },
  },
};